+++ Oqtane.Server/Controllers/ThemeController.cs
@@ -8,14 +8,14 @@ using System.Reflection;
 using System.Linq;
 using Microsoft.AspNetCore.Hosting;
 using Oqtane.Enums;
-using Oqtane.Infrastructure.Interfaces;
+using Oqtane.Infrastructure;
 using Oqtane.Repository;
 
 // ReSharper disable StringIndexOfIsCultureSpecific.1
 
 namespace Oqtane.Controllers
 {
-    [Route("{site}/api/[controller]")]
+    [Route("{alias}/api/[controller]")]
     public class ThemeController : Controller
     {
         private readonly IThemeRepository _themes;
@@ -54,34 +54,42 @@ namespace Oqtane.Controllers
         {
             List<Theme> themes = _themes.GetThemes().ToList();
             Theme theme = themes.Where(item => item.ThemeName == themename).FirstOrDefault();
-            if (theme != null)
+            if (theme != null && Utilities.GetAssemblyName(theme.ThemeName) != "Oqtane.Client")
             {
-                themename = theme.ThemeName.Substring(0, theme.ThemeName.IndexOf(","));
-
-                string folder = Path.Combine(_environment.WebRootPath, "Themes\\" + themename);
+                // clean up theme static resource folder
+                string folder = Path.Combine(_environment.WebRootPath, "Themes" , Utilities.GetTypeName(theme.ThemeName));
                 if (Directory.Exists(folder))
                 {
                     Directory.Delete(folder, true);
+                    _logger.Log(LogLevel.Information, this, LogFunction.Delete, "Theme Static Resources Removed For {ThemeName}", theme.ThemeName);
                 }
 
+                // remove theme assembly from /bin
                 string binfolder = Path.GetDirectoryName(Assembly.GetEntryAssembly().Location);
-                foreach (string file in Directory.EnumerateFiles(binfolder, themename + "*.dll"))
-                {
-                    System.IO.File.Delete(file);
-                }
-                _logger.Log(LogLevel.Information, this, LogFunction.Delete, "Theme Deleted {ThemeName}", themename);
+                System.IO.File.Delete(Path.Combine(binfolder, Utilities.GetAssemblyName(theme.ThemeName) + ".dll"));
+                _logger.Log(LogLevel.Information, this, LogFunction.Delete, "Theme Assembly {Filename} Removed For {ThemeName}", Utilities.GetAssemblyName(theme.ThemeName) + ".dll", themename);
 
                 _installationManager.RestartApplication();
             }
         }
 
-        // GET api/<controller>/load/filename
-        [HttpGet("load/{filename}")]
-        public IActionResult Load(string filename)
+        // GET api/<controller>/load/assembyname
+        [HttpGet("load/{assemblyname}")]
+        public IActionResult Load(string assemblyname)
         {
-            string binfolder = Path.GetDirectoryName(Assembly.GetEntryAssembly().Location);
-            byte[] file = System.IO.File.ReadAllBytes(Path.Combine(binfolder, filename));
-            return File(file, "application/octet-stream", filename);
+            if (Path.GetExtension(assemblyname).ToLower() == ".dll")
+            {
+                string binfolder = Path.GetDirectoryName(Assembly.GetEntryAssembly().Location);
+                byte[] file = System.IO.File.ReadAllBytes(Path.Combine(binfolder, assemblyname));
+                return File(file, "application/octet-stream", assemblyname);
+            }
+            else
+            {
+                _logger.Log(LogLevel.Error, this, LogFunction.Read, "User Not Authorized To Download Assembly {Assembly}", assemblyname);
+                HttpContext.Response.StatusCode = 401;
+                return null;
+            }
         }
+
     }
 }
